woziweidao_simulation_k8s-dev/requirements/dev.txt
flask == 1.1.4
flask-restx == 0.4.0  
werkzeug == 1.0.1
flask-cors == 3.0.10 
kafka-python==2.0.2
celery == 5.1.2
redis == 3.5.3
eventlet == 0.30.2
gunicorn == 20.1.0
pymysql == 1.0.2
flower==1.0.0
pandas==1.1.5
click == 7.1.2
SQLAlchemy==1.3.23
apscheduler == 3.7.0
gevent == 21.12.0